Note selection defined by FILTER arguments can now be printed, deleted, undeleted
