+++ note.go
@@ -2,14 +2,13 @@ package main
 
 import (
 	"errors"
-	"os"
+	"fmt"
+	"github.com/google/uuid"
 	"io"
 	"log"
-	"fmt"
-	"regexp"
-	_"time"
+	"os"
 	"path/filepath"
-	"github.com/google/uuid"
+	"regexp"
 )
